pyproject.toml
[project]
name = "sar-sr-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.5",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "rasterio>=1.5.0",
    "scikit-image>=0.26.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
